Add load regulation and backpressure
The service does not know how to handle the situation where the
HTTP/2 client cannot accept any more streams. This happens predictably
when connecting to APNS using JWT authentication. The APNS server sets
its maximum number of concurrent streams to 1 until authentication has
been done successfully.

In any case, it is not reasonable to expect that the APNS HTTP/2 server
will be able to accommodate an arbitrary number of concurrent streams,
so at some point the session needs to deal with "no more streams"
errors.

This commit makes the following improvements:

- A queue limit is added, to provide configurable backpressure. If the
  queue is full, requests are rejected with `{error, try_again_later)`.

- The concept of "submitted" no longer applies, because all requests are
  queued. No requests are sent directly to the HTTP/2 client.

- A configurable polling interval is added. The session polls the queue
  for notifications to send when the interval expires. When there are
  notifications to send, the session sends a burst of notifications to
  the HTTP/2 client. This is done in a blocking fashion, which is ok if
  the burst size is not too large and the polling interval not too
  small, because the session does an async HTTP/2 request. As long as
  the HTTP/2 concurrency size is reasonably large, throughput and
  responsiveness should be acceptable.

- A configurable maximum burst size is added, which limits the number of
  notifications that are dequeued and sent to the HTTP/2 client after
  each polling interval.

TODO
----
Right now, the burst mechanism sends notifications until it gets a
`REFUSED_STREAM` error or reaches `burst_max`. This is ok, but not
ideal. For example, `burst_max` might be smaller than the concurrent
stream limit set by the APNS server, which limits throughput
unnecessarliy.

It would be ideal if the load regulation kept track of the number of
active and maximum allowed concurrent HTTP/2 streams, and matched the
burst adaptively to keep the number of streams at the current maximum.

Unfortunately, there is no `on_settings_received` callback provided by
`chatterbox`, so the session will need to query the HTTP/2 client for
the `max_concurrent_streams` in `peer_settings` somehow.

Or add a callback to chatterbox; problem is that all the existing
callbacks are in `h2_stream.erl`, which does not handle connection
settings.

Another possibility is to write a simple stream callback module, and use
`on_request_end_stream/1` to dig into the callback state to get the
connection pid.

Or add `h2_connection:get_peer_settings/1`.