WIP: Use new invalidation API
We no longer delete APNS tokens when they are inactive; instead, we save
to database a timestamp we get from APNS when it returns an Unregistered
(410) error.

This functionality is provided by the
sc_push_reg_api:update_invalid_timestamp_by_svc_tok/2 API in sc_push_lib
v2.

This timestamp is in POSIX milliseconds, and is the most recent
time at which APNS marked the token as invalid. Evidently APNS checks
periodically to make sure the token has not been revived in the interim.

Client code can check if the last invalid timestamp is more recent than
the latest token registration. If so, the token is invalid and should
not be used for push notifications.

If the token is subsequently re-registered, for example, if someone
re-installs the same version of the deleted app on their iOS device and
gets the same token as before, and that token is stored in the database
with a new timestamp.  This will signal that the token is now active.